Убрал из Game игрока и перенес его в GameContext

+++ WarShooter/WarShooter/sheet.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "Assets.h"
 
 const sf::Vector2u WINDOW_SIZE = { 1500, 600 };
 const sf::Vector2f INITIAL_POSITION = { 400, 300 };
@@ -9,7 +10,7 @@ const sf::Vector2f BLOCK_POSITION2 = { 100, 200 };
 const sf::Vector2f BLOCK_POSITION3 = { 100, 350 };
 const sf::Vector2f BLOCK_SIZE = { 150, 50 };
 
-const sf::Vector2f BULLET_POSITION = { 450, 325 };
+const sf::Vector2i BULLET_POSITION = { 450, 325 };
 const sf::Vector2f BULLET_SIZE = { 10, 10 };
 
 struct KeyMap
